AdminTools V2.5 - The Rework, of the Rework
Drastically changed the plugin functionality to be more user supportive from V2

+++ AdminTools/Commands/Explode/Explode.cs
@@ -1,28 +1,29 @@
 ﻿using CommandSystem;
 using Exiled.API.Features;
 using Exiled.Permissions.Extensions;
+using Grenades;
 using System;
 
-namespace AdminTools.Commands.Kill
+namespace AdminTools.Commands.Explode
 {
     [CommandHandler(typeof(RemoteAdminCommandHandler))]
     [CommandHandler(typeof(GameConsoleCommandHandler))]
-    public class Kill : ParentCommand
+    public class Explode : ParentCommand
     {
-        public Kill() => LoadGeneratedCommands();
+        public Explode() => LoadGeneratedCommands();
 
-        public override string Command { get; } = "kill";
+        public override string Command { get; } = "expl";
 
-        public override string[] Aliases { get; } = new string[] { };
+        public override string[] Aliases { get; } = new string[] { "boom" };
 
-        public override string Description { get; } = "Kills everyone or a user instantly";
+        public override string Description { get; } = "Explodes a specified user or everyone instantly";
 
         public override void LoadGeneratedCommands() { }
 
         protected override bool ExecuteParent(ArraySegment<string> arguments, ICommandSender sender, out string response)
         {
             EventHandlers.LogCommandUsed((CommandSender)sender, EventHandlers.FormatArguments(arguments, 0));
-            if (!((CommandSender)sender).CheckPermission("at.kill"))
+            if (!((CommandSender)sender).CheckPermission("at.explode"))
             {
                 response = "You do not have permission to use this command";
                 return false;
@@ -30,7 +31,7 @@ namespace AdminTools.Commands.Kill
 
             if (arguments.Count != 1)
             {
-                response = "Usage: kill ((player id / name) or (all / *))";
+                response = "Usage: expl ((player id / name) or (all / *))";
                 return false;
             }
 
@@ -38,31 +39,45 @@ namespace AdminTools.Commands.Kill
             {
                 case "*":
                 case "all":
+                    if (arguments.Count != 1)
+                    {
+                        response = "Usage: expl (all / *)";
+                        return false;
+                    }
+
                     foreach (Player Ply in Player.List)
                     {
                         if (Ply.Role == RoleType.Spectator || Ply.Role == RoleType.None)
                             continue;
 
                         Ply.Kill();
+                        EventHandlers.SpawnGrenadeOnPlayer(Ply, GrenadeType.Frag, 0.1f);
                     }
-
-                    response = "Everyone has been game ended (killed) now";
+                    response = "Everyone exploded, Hubert cannot believe you have done this";
                     return true;
                 default:
+                    if (arguments.Count != 1)
+                    {
+                        response = "Usage: expl (player id / name)";
+                        return false;
+                    }
+
                     Player Pl = Player.Get(arguments.At(0));
                     if (Pl == null)
                     {
-                        response = $"Player not found: {arguments.At(0)}";
+                        response = $"Invalid target to explode: {arguments.At(0)}";
                         return false;
                     }
-                    else if (Pl.Role == RoleType.Spectator || Pl.Role == RoleType.None)
+
+                    if (Pl.Role == RoleType.Spectator || Pl.Role == RoleType.None)
                     {
-                        response = $"Player {Pl.Nickname} is not a valid class to kill";
+                        response = $"Player \"{Pl.Nickname}\" is not a valid class to explode";
                         return false;
                     }
 
                     Pl.Kill();
-                    response = $"Player {Pl.Nickname} has been game ended (killed) now";
+                    EventHandlers.SpawnGrenadeOnPlayer(Pl, GrenadeType.Frag, 0.1f);
+                    response = $"Player \"{Pl.Nickname}\" game ended (exploded)";
                     return true;
             }
         }
